# api/openapi.yaml
openapi: 3.1.0
info:
  title: omni-5e
  description: REST API for the D&D 5e System Reference Document
  version: 0.1.0
  license:
    name: CC-BY-4.0
    url: https://creativecommons.org/licenses/by/4.0/legalcode
  contact:
    name: omni-5e
    url: https://github.com/omni-5e/omni-5e
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
paths:
...
  /spells:
    get:
      summary: List spells
      operationId: listSpells
      tags: [spells]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: class
          in: query
          schema: { type: string }
          description: Filter by class slug
        - name: level
          in: query
          schema: { type: integer, minimum: 0, maximum: 9 }
        - name: school
          in: query
          schema: { type: string }
        - name: ritual
          in: query
          schema: { type: boolean }
        - name: concentration
          in: query
          schema: { type: boolean }
        - name: q
          in: query
          schema: { type: string }
          description: Full-text search
      responses:
        "200":
          description: Paginated list of spells
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpellList"
  /spells/{slug}:
    get:
      summary: Get a spell by slug
      operationId: getSpell
      tags: [spells]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Spell detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Spell"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    srd_version:
      name: srd_version
      in: query
      schema: { type: string }
      description: SRD content version (e.g. 5.2.1). Also via X-SRD-Version header.
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
    offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
    ProblemDetail:
      type: object
      properties:
        type: { type: string, example: "about:blank" }
        title: { type: string, example: "Not Found" }
        status: { type: integer, example: 404 }
        detail: { type: string }
...
    PaginationMeta:
      type: object
      properties:
        srd_version: { type: string }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    PaginationLinks:
      type: object
      properties:
        self: { type: string }
        next: { type: string, nullable: true }
        prev: { type: string, nullable: true }
    Spell:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        level: { type: integer }
        school: { type: string }
        casting_time: { type: string }
        range: { type: string }
        components:
          type: object
          properties:
            verbal: { type: boolean }
            somatic: { type: boolean }
            material: { type: boolean }
            material_detail: { type: string }
        duration: { type: string }
        concentration: { type: boolean }
        ritual: { type: boolean }
        description: { type: string }
        at_higher_levels: { type: string }
        class_slugs: { type: array, items: { type: string } }
    SpellList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Spell" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
...